fix(warnings): fix GL-format, null-deref and ignored-return warnings
- GL sysdep debug logs (ogl_core Oc*, tritium/core Ec*): cast uid_t/size_t
  args to (int) and Napi handles to (void*) so %d/%p match the argument type
  (clang -Wformat, gcc -Wformat=, MSVC C4477).
- LDOM2Tree::createObjByTypeNameT: drop a pScrObj->destruct() call inside the
  'pScrObj==NULL' branch. destruct() is pure virtual, so this is a latent null
  dereference; only an error path (malformed type name) could reach it, which
  is why it never crashed in practice (gcc -Wnonnull).
- PosixProcImpl/CnsParFile/test_msglog: check the return of chdir/fgets/fread
  instead of discarding it (gcc -Wunused-result).

+++ src/sysdep/ogl_core/OcBufTexRep.cpp
@@ -27,20 +27,20 @@ void OcBufTexRep::init(gfx::DisplayContext *pdc)
     m_nTexID = texid;
     CHK_GLERROR("OcBufTexRep::init glGenTextures");
 
-    MB_DPRINTLN("OcBufTexRep::init view=%d, buf=%d, tex=%d", m_nViewID, m_nBufID, m_nTexID);
+    MB_DPRINTLN("OcBufTexRep::init view=%d, buf=%d, tex=%d", (int)m_nViewID, m_nBufID, m_nTexID);
 }
 
 OcBufTexRep::~OcBufTexRep()
 {
     MB_DPRINTLN("OcBufTexRep::~OcBufTexRep view=%d, buf=%d, tex=%d",
-                m_nViewID, m_nBufID, m_nTexID);
+                (int)m_nViewID, m_nBufID, m_nTexID);
 
     if (m_nBufID == 0 && m_nTexID == 0) return;
 
     qsys::ViewPtr rvw = qsys::SceneManager::getViewS(m_nViewID);
     if (rvw.isnull()) {
         MB_DPRINTLN("OcBufTexRep> unknown parent view (%d), buf/tex cannot be deleted",
-                    m_nViewID);
+                    (int)m_nViewID);
         return;
     }
